Changed packet loss to 20% and fixed resend bug

+++ ZRTP/stateengine.h
@@ -66,7 +66,7 @@ private:
     Event *actualEvent;
 
     uint8_t *sentMessage;
-    uint8_t sentMessageLength;
+    uint32_t sentMessageLength;
 
     zrtpTimer T1;
     zrtpTimer T2;
